Data
In this project, we are working with a dataset containing images of both benign and malignant cancerous cells. These images are crucial for training and testing a deep learning model to classify cancer types accurately.

Training Images: A collection of images representing both benign and malignant tumors, used to train the model. These images allow the model to learn the distinct features of each class.
Testing Images: A separate set of images, not seen by the model during training, used to evaluate the performance and generalization of the model. This helps assess how well the model can classify new, unseen images.
The dataset includes a variety of tumor images, and it is essential for building a model capable of distinguishing between good and bad (benign and malignant) cases with high accuracy. These images are labeled and will be used in both the training and testing phases of the project.

The images are provided in various formats and sizes to ensure the model is trained effectively and can perform classification tasks on real-world data.