Data/Train/Malignant
These images represent malignant (cancerous) tumors and are used in the training phase of the project. The model will learn to identify the features that characterize malignant tumors, helping it differentiate between benign and malignant cases during testing. This dataset is essential for training the model to recognize cancerous conditions accurately.